requirements.txt
slack-bolt
python-dotenv
openai
ultra_rest_client
slackstyler
requests